requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.2
google-generativeai==0.3.0
slowapi==0.1.9
pydantic==2.0.0